requirements.txt
Flask_MySQLdb==0.2.0
Flask_Bcrypt==0.7.1
Flask==2.0.1
Flask_Mail==0.9.1
hdbcli
python-dotenv
